Massive rewrite of the parenting system. getParent is now deprecated (though it won't issue warnings). Each object has a tree of parent objects (currently) retrieved dynamically by using only one weakref to the object that directly owns the sub-object.
Additionally, and the reason the above changes were made, objects now have a built in self-observation system. These self-observations are removed from the dispatcher whenever the object goes away. This should hopefully resolve a real/theoretical problem of the dispatcher overflowing with dead objects.

This is all a bit slow. Speed is the next step.