Add default values as it was indicated in the pre refactored version.
The old version had for some values a list as a default value instead
of None, which is the most common default. However, the setters Didn't
reset the attributes to lists, instead they where reset to None.

Now setting None will set the default. It has been taken care of not using
the same mutable objects when setting the default value, using copy.copy.
Thus, a consumer can't change the value of a default.