Remove kerning validation during read, save and reload.
For the record, I am opposed to this change as it allows the defcon
kerning object to become populated with conflicting kerning pairs. This
may very well destabilize defcon's internal kerning handling. It also
allows defcon to write conflicting kerning pairs. At some point in the
journey towards a compiled font this conflict is going to result in the
user being presented with an error that they must solve. By loosening
these validation restrictions, we are making it significantly harder
for the user to figure out what happened to their kerning and which
environment is responsible for creating the conflicting data. That is
going to make it harder for the user to figure out how to correct the
problem. This could be avoided by environments following the logic and
not writing conflicting data. However, that seems to be an undesirable
solution so we're going to shift the responsibility for creating valid
data structures to the user. I do not agree with this, but my opinion
is in the minority.

See the discussion in issue #38 for a summary of the issue.