add support for UFOZ and fs-powered ufoLib
This patch enables loading/saving a defcon.Font from/to zipped UFO
(*.ufoz) using the latest ufoLib (yet unreleased, master branch).

The save method has a new 'structure' keyword argument, which can
be either "zip" or "package" (or a value from UFOFileStructure enum).
The Font has a new 'ufoFileStructure' read-only attribute that
stores the original file structure from the UFOReader (or None if
the font is created from scratch).

The rest of the patch is basically avoiding to use os.path interface,
e.g. path separator is always "/" with pyfilesystem2.

I marked a few of the problematic spots with 'XXX'.