only accept strings or os.PathLike as 'path'
all of defcon API assumes that font.path is a path string.
It's probably better to keep it that way.
Earlier I was experimenting with 'path' being either a string or an
already instantiated fs.base.FS object.
But I changed my mind, i think it confuses things.
Let's keep it simple for now. Path must be a string to a local
path, either a folder or a zipfile.
We'll experiment with FS objects in the fontTools.ufoLib when we get there